Comisiones

+++ PAR_INTRANET/Models/Clases/Tablas/RefArt.cs
@@ -2,8 +2,6 @@
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using System.Linq;
-using System.Web;
 
 namespace PAR_INTRANET.Models.Clases
 {
@@ -11,7 +9,7 @@ namespace PAR_INTRANET.Models.Clases
     public class RefArt
     {
         [Key]
-        public Byte CodRefArt { get; set; }
+        public int CodRefArt { get; set; }
 
         [StringLength(30)]
         public string DesRefArt { get; set; }
